// frontend/src/app/enums/baseURL.ts
enum BaseURL {
    // FRONTEND URL
    FE_PRODUCTION = "https://greensync.com.lk",
    FE_DEVELOPMENT = "https://dev.greensync.com.lk",
    FE_LOCALHOST = "http://localhost:4200",
    // BACKEND URL
    BE_DEVELOPMENT = 'https://backend-dev.greensync.com.lk',
    BE_PRODUCTION = 'https://backend.greensync.com.lk',
    BE_LOCALHOST = 'http://localhost:5001'
}

export default BaseURL;